Get the inferred regression coefficients

Get the samples from the posterior distribution of the regression coefficient matrix B, together with the posterior mean and quantiles. The regression coefficient matrix B is a matrix where the number of rows is defined by the number of traits that are modelled, and the number of columns is the number of columns of the matrix m$X (the number of explanatory variables after transformation via formula)

getB(m)

Arguments

m

a model fitted with jtdm_fit

Value

A list containing:

Bsamples

Sample from the posterior distribution of the regression coefficient matrix. It is an array where the first dimension is the number of traits, the second the number of columns in m$X (the number of variables after transformation via formula) and the third the number of MCMC samples.

Bmean

Posterior mean of the regression coefficient matrix.

Bq975,Bq025

97.5% and 0.25% posterior quantiles of the regression coefficient matrix.

Examples

data(Y)  
data(X)
# Short MCMC to obtain a fast example: results are unreliable !
m = jtdm_fit(Y=Y, X=X, formula=as.formula("~GDD+FDD+forest"),  adapt = 10, 
        burnin = 100, 
        sample = 100) 
#> Compiling rjags model...
#> Calling the simulation using the rjags method...
#> Note: the model did not require adaptation
#> Burning in the model for 100 iterations...
#> Running the model for 100 iterations...
#> Extending 100 iterations for pD/DIC estimates...
#> Simulation complete
#> Calculating summary statistics...
#> Calculating the Gelman-Rubin statistic for 21 variables....
#> Note: Unable to calculate the multivariate psrf
#> Finished running the simulation
# get the inferred regression coefficients
B=getB(m)